pipeline/consumer/requirements.txt
kafka-python==2.0.2
clickhouse-connect==0.8.9
six==1.16.0